Extracted some logic out of the eventRepository for getting a trace. This will be used on the frontend
